added outline for linear disassemble and cleaned traits
added in a rough outline for the linear disassemble, it increases coverage a lot but it's too slow, and it adds too many extra invalid bb.

I need to clean it up with some filtering before enabling it.

Also added some filtering to the trait extractor to filter out invalid bb